+++ src/CMakeLists.txt
@@ -14,7 +14,7 @@ add_executable(kernel.elf
         kernel/boot/idt/interrupt_handler.asm
         kernel/boot/idt/exception_handler.asm
         kernel/boot/gdt/global_descriptor_table.asm
-        kernel/boot/tss/tss.asm
+        kernel/boot/tss/task_state_segment.asm
         kernel/terminal/Terminal.cpp
         kernel/drivers/video_buffer_display.cpp
         kernel/boot/init.cpp
@@ -23,7 +23,7 @@ add_executable(kernel.elf
         kernel/boot/idt/interrupt_descriptor_table.cpp
         kernel/boot/idt/interrupt_handler.cpp
         kernel/boot/idt/programmable_interrupt_controller.cpp
-)
+        kernel/boot/tss/task_state_segment.cpp kernel/boot/gdt/global_descriptor.cpp)
 
 target_link_libraries(kernel.elf PRIVATE project_options) # project_warnings)
 target_include_directories(kernel.elf PUBLIC .)
