+++ src/kernel/boot/tss/task_state_segment.asm
@@ -1,8 +1,8 @@
 
 
-global initializeTaskStateSegment
+global loadTaskRegister
 
-initializeTaskStateSegment:
+loadTaskRegister:
 	            mov             ax, (5 * 8) | 0                 ; fifth 8-byte selector, symbolically OR-ed with 0 to set the RPL (requested privilege level).
 	            ltr             ax
 	            ret
