+++ src/kernel/kernel.cpp
@@ -18,7 +18,7 @@
 #include <kernel/drivers/video_buffer_display.h>
 #include "terminal/Terminal.h"
 #include "lib/libc/stdint.h"
-#include "boot/tss/tss.h"
+#include "kernel/boot/tss/task_state_segment.h"
 
 namespace kernel {
